crypto_bot/requirements.txt
python-telegram-bot[job-queue]==21.6
groq==0.11.0
aiohttp>=3.10.5
sqlalchemy==2.0.35
flask==3.0.3
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2
cachetools==5.5.0
